Increase Gradle JVM memory limit and map preview quality
The Gradle JVM memory limit has been increased to 4096 MB and the MaxMetaspaceSize to 1024 MB. These changes are reflected in the 'android/gradle.properties' file.

The quality of the map preview image generated by the Mapbox API has been improved by increasing the zoom level from 14 to a variable zoom level set in the 'util/location.js' file. This means that users will now see a more detailed preview of the selected location when picking or viewing a location on the map.