Update LocationPicker and PlaceDetails to use zoomLevel instead of zoom.
This commit updates the LocationPicker and PlaceDetails components to use the 'zoomLevel' property instead of 'zoom', which is more accurate and consistent with other parts of the app. Specifically, in LocationPicker, the 'initialZoomLevel' property is updated to use 'pickedLocation.zoomLevel' instead of 'pickedLocation.zoom'. In PlaceDetails, the 'initialZoomLevel' property is added with the value of 'fetchedPlace.location.zoomLevel'.