Fix typo in Info screen and update state variable names in Map screen
In the Info screen, a typo in a text message was corrected to reflect accurate information about location data being used in Slovenia and Austria.

In the Map screen, state variable names were updated for clarity and consistency. 'location' was changed to 'currentLocation' to better reflect that it represents the current GPS location of the user. 'zoomLevel' was changed to 'currentZoomLevel' to match the updated name for the supporting state variable. Additionally, a delay before setting the current location to null was increased from 2 to 3 seconds to allow enough time for the user to see the current location icon on the map before it disappears.