Increase gradle memory limit and update app version
The gradle memory limit was increased from 2048m to 4096m and the maximum metaspace size was increased from 512m to 1024m. The app version was also updated from 1.0.0 to 1.1.0. In addition, the default zoom level in the Map screen was changed from 14 to 8.1 and the default location was changed to show the whole municipality where the Ledinska imena are located.